requirements.txt
Flask
gTTS
pydub
speechrecognition
langdetect
python-dotenv
google-generativeai
